+++ packages/mparty/src/interfaces/IMparty.ts
@@ -1,15 +1,22 @@
-import { IAdapter } from './IAdapter';
-import { IMpartyLimits } from './IMpartyValidator';
+import { IncomingMessage } from 'http';
 
-export interface IMpartyOptions {
+import { IAdapter, IFileMetadata } from './IAdapter';
+import { IMpartyLimits } from './IMpartyLimits';
+
+export interface IMpartyOptions<T extends IFileMetadata> {
   /** Adapter to be used for file upload */
-  adapter?: IAdapter;
+  // adapter?: IAdapter<T> | ((req: IncomingMessage) => IAdapter<T>);
+  adapter?: IAdapter<T>;
 
   /** If no adapter provided and provided a destionation - FsAdapter will be used for provided destionation */
+  // destination?: string | ((req: IncomingMessage) => string);
   destination?: string;
 
   /** Whether to throw an error on requests with application/json Content-Type. Default: false  */
-  shouldFailOnJson?: boolean;
+  failOnJson?: boolean;
+
+  /** Whether to remove uploaded files from storage on Error occured. Default: true */
+  removeUploadedFilesOnError?: boolean;
 
   /**
    *  Busboy option. If paths in the multipart 'filename' field shall be preserved. (Default: false).
@@ -27,6 +34,6 @@ export interface IMpartyOptions {
 }
 
 /* eslint-disable-next-line */
-export interface IMpartyUploadOptions extends IMpartyOptions {
+export interface IMpartyUploadOptions<T extends IFileMetadata> extends IMpartyOptions<T> {
   //
 }