+++ packages/mparty/src/interfaces/index.ts
@@ -3,6 +3,7 @@ import '@tsrt/types';
 export * from './AdapterEventEmitter';
 export * from './IMparty';
 export * from './IMpartyValidator';
+export * from './IMpartyLimits';
 export * from './IBusboy';
 export * from './IAdapter';
 export * from './IFsAdapter';