+++ packages/mparty/src/utils/helpers.ts
@@ -1,10 +1,10 @@
-import { isEmpty } from '@tsrt/utils';
+import { isEmpty, getRandomString } from '@tsrt/utils';
 
 export { getFileExtension } from '@tsrt/utils';
 
 export function createFileName(originalFileName: string): string {
   const timestamp = Date.now();
-  return `${timestamp}_${originalFileName}`;
+  return `${timestamp}_${getRandomString()}_${originalFileName}`;
 }
 
 export function getFirstItem<T extends GenericObject>(object: T): T[keyof T] {