+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.0
+
+### New features
+
+- [@tsrt/mparty] Added framework agnostic package for handling multipart/form-data requests based on busboy
+- [@tsrt/mparty] Added default FileSystem Adapter
+- [@tsrt/mparty] Added files and fields basic validations
+
+### 
+
+---
+
 ## 0.1.2
 
 ### New features